git: fix handling of git repos using master as version
Since changing to git init followed by fetch to initialise
git cache, when using master branch as the target version
for a package, git refuses to fetch in to currently checkout
out branch or master.

git init always creates a master branch, so this will always
be an issue for any package targeting master.
Add -u option to git fetch command to allow it to ignore that
check. We know that it is okay to overwrite master as we know that
it either contains nothing after the initial git init command, or
is tracking the remote master after a successful fetch.
We always want it to overwrite it.

Signed-off-by: Robert Beckett <bbeckett@netvu.org.uk>
Origin: http://lists.buildroot.org/pipermail/buildroot/2018-June/222961.html
Issue: #MCOM03SW-122
Change-Id: I54f463f9541237e40068d692e6a9836a9c04fc7d