python-serial: Revert to v3.1
The recipe for building python-serial incorrectly uses
distutils instead of setuptools due to an incorrect merge
of tag 2018.02 into branch mcom02.
This causes confusing build failures, for example, when
systemd is enabled.

This commit reverts python-serial to the version used in
Buildroot 2018.02.3. As a result, python-serial.mk becomes
identical to its upstream counterpart.

Issue: #MCOM02SW-916
Change-Id: Id31f303d4972f49ce28c842da71ec62198b0ab26