linux: Fix edge cases of Device Tree wildcard expansion
GNU Make wildcards are different from shell wildcards,
in that they omit patterns entirely if no matches are found.
Shell wildcards leave the pattern unchanged if nothing was found.
The latter behavior is more preferable in our case, since it
causes DT compilation to error out if the pattern was specified wrong,
and not just silently ignore missing files.

Because of this, it was not noticed that DT builds didn't run
without wildcards, since the .dts suffix wasn't appended to them.

This commit fixes both issues.

Change-Id: If0b9799d9806d325abd0b2467a8ebb78bbb3adc6