package/flatbuffers: fix build errors
When updating MCom-03 Buildroot to 2023.05, the git history of
the flatbuffers package was incorrectly applied.

This is related to the fact that earlier commits were cherry-picked
from upstream, skipping the "package/flatbuffers: bump version to 2.0.8"
commit for building the tensorflow-lite v2.12.0 package.

There is no way to restore the git history for the flatbuffers package.
MCom-03 Buildroot does not use musl, so we are removing the patch to support it.

In the presence of a patch, the build fails with an error:

  Applying 0001-include-flatbuffers-base.h-fix-build-on-musl.patch using patch:
  patching file include/flatbuffers/base.h
  Hunk #1 FAILED at 264.
  1 out of 1 hunk FAILED -- saving rejects to file include/flatbuffers/base.h.rej

Also fixed an error with the `make legal-info` command:

  >>> flatbuffers 2.0.6 Collecting legal info
  ERROR: No hash found for LICENSE

Issue: #MCOM03SW-2716
Change-Id: I7f7b82336139439cc26aa96963cea63001a62c06