package/uboot-tools: Add missing host-util-linux dependency
Upstream Buildroot 2023.05 (we are based on) has
package/uboot-tools/uboot-tools.mk tuned for U-Boot 2021.07 and
we use U-Boot 2023.04.

To build uboot-tools with mkeficapsule support, we need the uuid
header, which is obtained from util-linux.

Missing dependency is quickly detected when Buildroot is built
in parallel with log:

  tools/mkeficapsule.c:18:10: fatal error: uuid/uuid.h: No such file or directory
     18 | #include <uuid/uuid.h>
        |          ^~~~~~~~~~~~~
  compilation terminated.
  make[3]: *** [scripts/Makefile.host:95: tools/mkeficapsule] Error 1

Issue: #MCOM03SW-2720
Change-Id: I2614a419e664750c7cb55c61fdecf21a5f5da6ac