uboot-tools: Support U-Boot 2024.07
Similarly to fw_env, printinitialenv uses config.h to define the default
environment of the created image, so it isn't necessary for u-boot tools
which are built without a board configuration. Drop its inclusion to
avoid a build error:

  tools/printinitialenv.c:20:10: fatal error: config.h: No such file or
  directory
     20 | #include <config.h>

Patch 0002-tools-only-in-no-dot-config-targets enables a "tools-only"
build for uboot-tools. Resolve a conflict in the patch after support
for building u_boot_tools PyPi package was added in commit 75554dfac298
("patman: Add support for building a u_boot_tools PyPi package"):

  Applying 0002-tools-only-in-no-dot-config-targets.patch using patch:
  patching file Makefile
  Hunk #1 FAILED at 522.

Add "Upstream: N/A" line to pass check-package as we don't know
whether a patch was ever sent to Buildroot.

Issue: #MCOM03SW-2940
Change-Id: I898474d7b0abf60ca75fda9af2fdbbc8c696b5d9