package/alsa-utils: Fix install rule for alsactl
The installation rule for systemd services does not create dependency
links in /usr/lib/systemd/system/sound.target.wants.
This means that the order in which alsa-restore and alsa-state services
are started is not defined, and it is possible that services will start
before the sound.target is reached.

Issue: #MCOM03SW-3747
Change-Id: I3df9f3bf6600e68d72f84403d3892088ef929085